requirements.txt
aenum==3.1.11
ipykernel==6.13.0
ipyleaflet==0.16.0
ipywidgets==7.7.0
kaleido==0.2.1
matplotlib==3.5.2
notebook==6.4.11
numpy==1.22.4
opencv-python==4.5.5.64
pandas==1.4.2
Pillow==9.1.1
plotly==5.8.0
pycocotools==2.0.4
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Shapely==1.8.2
statsmodels==0.13.2
tabulate==0.8.9
tqdm==4.64.0
xmltodict==0.13.0